pyproject.toml
[tool.poetry]
name = "shantaagent"
version = "0.1.0"
description = "building a chatbot using lanchain"
authors = ["Prashant Malge <prashantmalge181@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Framework :: FastAPI",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
duckduckgo-search = ">=7.1.1"
fastapi = "^0.115.5"
httpx = "^0.27.2"
langchain-core = "^0.3.20"
langchain-community = "^0.3.7"
langchain-openai = "^0.2.9"
langchain-anthropic = "^0.3.0"
langchain-google-genai = "^2.0.5"
langchain-groq = "^0.2.1"
langchain-aws = "^0.2.7"
langgraph = "^0.2.53"
langgraph-checkpoint-sqlite = "^2.0.1"
langsmith = "^0.1.145"
numexpr = "^2.10.1"
pyarrow = ">=18.1.0"
pydantic = "^2.10.1"
pydantic-settings = "^2.6.1"
pyowm = "^3.3.0"
python-dotenv = "^1.0.1"
setuptools = "^75.6.0"
streamlit = "^1.40.1"
tiktoken = ">=0.8.0"
uvicorn = "^0.32.1"
prometheus-client = ">=0.21.1,<0.22.0"
psutil = ">=6.1.1,<7.0.0"
opik = "^1.4.2"

[tool.poetry.group.dev.dependencies]
pre-commit = "*"
pytest = "*"
pytest-cov = "*"
pytest-env = "*"
pytest-asyncio = "*"
ruff = "*"

[tool.poetry.group.client.dependencies]
httpx = "^0.27.2"
pydantic = "^2.10.1"
python-dotenv = "^1.0.1"
streamlit = "^1.40.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
extend-select = ["I", "U"]

[tool.pytest.ini_options]
pythonpath = ["src"]
asyncio_default_fixture_loop_scope = "function"

[tool.pytest_env]
GROQ_API_KEY = "sk-fake-openai-key"